refactor(flow): lift parseJsonField + sanitizeJobResultForPublic into @sentropic/flow/job-helpers (BR26 Lot 7 real reorg slice 2)
